libmelo: playlist: add two callbacks to get status
This patch adds two new callbacks to MeloPlaylist, named has_prev and
has_next. It allows to know if a media is available in playlist
respectively before and/or after current media.
It also adds support of these callbacks into MeloPlaylistSimple.
Finally, the JSON-RPC method player.get_status is updated to handle
the playlist status.