libmelo: player: add new callback for media load
This patch adds a new callback and its access function to MeloPlayer:
load() allows user to load a media in paused or stopped state. It can
be very useful when media need to be loaded in player without playing
it instantly.
It also adds two new states to MeloPlayerState for loading and
buffering states when player is paused.