libmelo: playlist: simple: update player status
In order to maintain up-to-date the player status regarding playlist
status (has a previous or next media available in playlist), a new
function is added to MeloPlaylistSimple, which updates player status
on every changes into playlist.
A fix is also done on get_prev() which was not using the 'set'
parameter.