lib: add MeloSink for global audio sink management
A new class named MeloSink is added to Melo for global audio sinks
management. Thanks to this class an audio mixer can be implemented
with which all audio streams can be controlled in once: volume and
mute of each MeloPlayer using MeloSink instances can be controlled
through this new interface. It also allows to setup a general
configuration for all audio sinks (samplerate, channels) and to
control global volume / mute.

This patch also adds three JSON-RPC methods named 'sink.get_list',
'sink.get' and 'sink.set' respectively for getting list of all sinks,
for getting status of a specific sink and for setting volume / mute of
a specific sink.